requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.13.1
frozenlist==1.4.1
fsspec==2024.2.0
idna==3.7
importlib_metadata==7.1.0
intel-openmp==2021.4.0
ipykernel==6.29.4
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
lightning==2.2.5
lightning-utilities==0.11.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==10.2.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.5
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
stack-data==0.6.3
sympy==1.12
tbb==2021.11.0
threadpoolctl==3.5.0
torch==2.3.1+cu118
torchaudio==2.3.1+cu118
torchmetrics==1.4.0.post0
torchvision==0.18.1+cu118
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.9.0
tzdata==2024.1
wcwidth==0.2.13
yarl==1.9.4
zipp==3.19.2
